add review , comment view your reviews in progress ( user account)

+++ mylibrary/src/main/java/com/alina/mylibrary/model/db/Book.java
@@ -64,7 +64,7 @@ public class Book {
     private int numberOfReviews;
 
     @Column
-    private float bookRating;
+    private Double bookRating;
 
     @Column
     @NotNull
